Rename classes Memo -> LazySelfie.
