+++ jvm/selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/LazySelfie.kt
@@ -13,30 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.selfie.coroutines
+package com.diffplug.selfie
 
-import com.diffplug.selfie.Mode
-import com.diffplug.selfie.Roundtrip
-import com.diffplug.selfie.Selfie
-import com.diffplug.selfie.Snapshot
 import com.diffplug.selfie.guts.DiskStorage
 import com.diffplug.selfie.guts.LiteralString
 import com.diffplug.selfie.guts.LiteralValue
 import com.diffplug.selfie.guts.TodoStub
 import com.diffplug.selfie.guts.recordCall
 
-class MemoStringSuspend<T>(
+class LazySelfie<T>(
     private val disk: DiskStorage,
     private val roundtrip: Roundtrip<T, String>,
-    private val generator: suspend () -> T
+    private val generator: () -> T
 ) {
-  suspend fun toMatchDisk(sub: String = ""): T {
+  fun toMatchDisk(sub: String = ""): T {
     return toMatchDiskImpl(sub, false)
   }
-  suspend fun toMatchDisk_TODO(sub: String = ""): T {
+  fun toMatchDisk_TODO(sub: String = ""): T {
     return toMatchDiskImpl(sub, true)
   }
-  private suspend fun toMatchDiskImpl(sub: String, isTodo: Boolean): T {
+  private fun toMatchDiskImpl(sub: String, isTodo: Boolean): T {
     val call = recordCall(false)
     if (Selfie.system.mode.canWrite(isTodo, call, Selfie.system)) {
       val actual = generator()
@@ -61,13 +57,13 @@ class MemoStringSuspend<T>(
       }
     }
   }
-  suspend fun toBe_TODO(unusedArg: Any? = null): T {
+  fun toBe_TODO(unusedArg: Any? = null): T {
     return toBeImpl(null)
   }
-  suspend fun toBe(expected: String): T {
+  fun toBe(expected: String): T {
     return toBeImpl(expected)
   }
-  private suspend fun toBeImpl(snapshot: String?): T {
+  private fun toBeImpl(snapshot: String?): T {
     val call = recordCall(false)
     val writable = Selfie.system.mode.canWrite(snapshot == null, call, Selfie.system)
     if (writable) {
