+++ jvm/selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/coroutines/LazySelfieBinarySuspend.kt
@@ -27,7 +27,7 @@ import com.diffplug.selfie.guts.recordCall
 import kotlin.io.encoding.Base64
 import kotlin.io.encoding.ExperimentalEncodingApi
 
-class MemoBinarySuspend<T>(
+class LazySelfieBinarySuspend<T>(
     private val disk: DiskStorage,
     private val roundtrip: Roundtrip<T, ByteArray>,
     private val generator: suspend () -> T
